+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Clippy and Formatting updates
 - Submitted updates for `invert` and static color defines in SDL2 (PR #965 in Rust-SDL2)
 - Created text helper function (render_text) (#276)
+- Created Tab Bar Widget (#50)
 
 ## 0.4.26